hf_space/requirements.txt
streamlit
torch
transformers>=4.46.3
tokenizers>=0.20.3
einops
addict
easydict
Pillow
pytesseract
tqdm
numpy
